+++ Plugins/FreeRDP/Viewer/ParameterFreeRDP.cpp
@@ -2,12 +2,14 @@
 #include <QSettings>
 
 CParameterFreeRDP::CParameterFreeRDP(QObject *parent) : CParameter(parent),
-    m_nReconnectInterval(5)
+    m_nReconnectInterval(5),
+    m_bShowVerifyDiaglog(true)
 {}
 
 int CParameterFreeRDP::Load(QSettings &set)
 {
     CParameter::Load(set);
+    
     QString szDomain = set.value("FreeRDP/Domain").toString();
     freerdp_settings_set_string(
                 m_pSettings, FreeRDP_Domain, szDomain.toStdString().c_str());
@@ -24,12 +26,16 @@ int CParameterFreeRDP::Load(QSettings &set)
 
     SetReconnectInterval(set.value("FreeRDP/ReconnectionInterval",
                                    GetReconnectInterval()).toInt());
+    
+    SetShowVerifyDiaglog(set.value("FreeRDP/ShowVerifyDiaglog",
+                                   GetShowVerifyDiaglog()).toBool());
     return 0;    
 }
 
 int CParameterFreeRDP::Save(QSettings &set)
 {
     CParameter::Save(set);
+    
     set.setValue("FreeRDP/Domain", freerdp_settings_get_string(
                      m_pSettings, FreeRDP_Domain));
     set.setValue("FreeRDP/Width", freerdp_settings_get_uint32(
@@ -38,7 +44,9 @@ int CParameterFreeRDP::Save(QSettings &set)
                      m_pSettings, FreeRDP_DesktopHeight));
     set.setValue("FreeRDP/ColorDepth", freerdp_settings_get_uint32(
                      m_pSettings, FreeRDP_ColorDepth));
+    
     set.setValue("FreeRDP/ReconnectionInterval", GetReconnectInterval());
+    set.setValue("FreeRDP/ShowVerifyDiaglog", GetShowVerifyDiaglog());
     return 0;
 }
 
@@ -60,4 +68,14 @@ void CParameterFreeRDP::SetReconnectInterval(UINT newReconnectInterval)
         freerdp_settings_set_bool(m_pSettings,
                                   FreeRDP_AutoReconnectionEnabled, false);
     emit sigReconnectIntervalChanged();
+}
+
+bool CParameterFreeRDP::GetShowVerifyDiaglog() const
+{
+    return m_bShowVerifyDiaglog;
+}
+
+void CParameterFreeRDP::SetShowVerifyDiaglog(bool bShow)
+{
+    m_bShowVerifyDiaglog = bShow;
 }